requirements.txt
asgiref==3.2.10
Django==3.1.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-globals==0.3.2
djangorestframework==3.12.1
importlib-metadata==2.0.0
Markdown==3.2.2
Pillow==7.2.0
pytz==2020.1
PyYAML==5.3.1
sqlparse==0.3.1
zipp==3.3.0 